Changed RobotSpecs to use float data type

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/devices/RobotSpecs.java
@@ -8,11 +8,11 @@ package org.firstinspires.ftc.teamcode.devices;
 public class RobotSpecs {
 
     // TODO: Update Parameters of Robot
-    public static final double
-            robotHeight = 336.55,
-            robotWidth = 336.55,
-            cameraHeightFromGround = 334.55,
-            cameraAngle = 15.4;
+    public static final float
+            robotHeight = 336.55F,
+            robotWidth = 336.55F,
+            cameraHeightFromGround = 334.55F,
+            cameraAngle = 15.4F;
 
 
 }
